refactor: split db/src/progress.rs into submodules (#1952)
db/src/progress.rs crossed the ~800-line file-shape cap. Split into
db/src/progress/{mod,upsert,state,resume,session}.rs by concern, no
behavior change: upsert.rs holds write-path validation, state.rs the
per-book position/statistics/rate read-write, resume.rs the resume-card
aggregation, session.rs the session-report recording.
