+++ db/src/progress.rs
@@ -3,14 +3,29 @@
 //! `(user_id, book_uuid, format)` by client event time, not server receipt
 //! time. Session inserts go to the per-format `reading_sessions` /
 //! `listening_sessions` tables; all rows soft-reference `books.uuid`.
+//!
+//! Split by concern: [`upsert`] holds the write-path validation, [`state`]
+//! the per-book position/statistics/rate read-write, [`resume`] the
+//! resume-card aggregation, and [`session`] the session-report recording.
+//! Re-exported here as `omnibus_db::progress::*`.
 
-use omnibus_shared::{
-    AudiobookPlaybackRateRecord, AudiobookPlaybackRateUpdate, ChapterInfo, ProgressFormat,
-    ProgressRecord, ProgressUpdate, ResumePoint, SessionReport,
+use omnibus_shared::ProgressFormat;
+
+mod resume;
+mod session;
+mod state;
+mod upsert;
+
+#[cfg(test)]
+mod tests;
+
+pub use resume::{recent_progress, resume_points};
+pub use session::{insert_session_tx, record_session, record_session_tx};
+pub use state::{
+    attach_derived_kobo_location, get_playback_rate, get_progress, set_kobo_statistics_tx,
+    set_playback_rate, KoboStatistics,
 };
-use sqlx::{Row, Sqlite, SqlitePool, Transaction};
-
-use crate::{hls, resolve_canonical_book_uuid, resolve_canonical_book_uuid_exec};
+pub use upsert::{upsert_progress, upsert_progress_tx};
 
 #[derive(Debug, thiserror::Error)]
 pub enum ProgressError {
@@ -20,10 +35,10 @@ pub enum ProgressError {
     Sqlx(#[from] sqlx::Error),
 }
 
-impl From<hls::HlsError> for ProgressError {
-    fn from(e: hls::HlsError) -> Self {
+impl From<crate::hls::HlsError> for ProgressError {
+    fn from(e: crate::hls::HlsError) -> Self {
         match e {
-            hls::HlsError::Db(inner) => ProgressError::Sqlx(inner),
+            crate::hls::HlsError::Db(inner) => ProgressError::Sqlx(inner),
         }
     }
 }
@@ -60,743 +75,3 @@ fn parse_format(raw: &str) -> ProgressFormat {
         _ => ProgressFormat::Epub,
     }
 }
-
-/// Backward jump in an accepted audio position write that's worth a WARN
-/// (~10 minutes). A normal skip-back or chapter re-listen is well under
-/// this; the 2026-08-11 chapter-19-to-13 revert was hours.
-const AUDIO_BACKWARD_JUMP_THRESHOLD_SECONDS: f64 = 600.0;
-
-/// Upsert a position row for `(user, book, format)` and return the
-/// **surviving** server-authoritative record. Resolves the request uuid to
-/// the **canonical** `books.uuid` (keeping the `BookNotFound` guard — you
-/// cannot record progress for a book the server has never indexed) and
-/// stores/keys on it.
-///
-/// The conflict resolution is conditional on `client_updated_at`, not
-/// unconditional last-write-wins on receipt order: the stored
-/// `MIN(update.client_updated_at, now)` — clamping a fast client clock so it
-/// can't pin itself as permanently newest — only overwrites the row when it
-/// is `>=` the value already stored. A write with no `client_updated_at`
-/// (older client) is treated as "now", matching plain last-write-wins. A
-/// rejected (older) write still re-reads and returns the row that won, so
-/// the caller learns it is behind rather than seeing its own rejected
-/// payload echoed back.
-///
-/// An accepted write replaces the **whole position** — `epub_cfi`,
-/// `progress_percent`, and `kobo_location` are one atomic snapshot, never
-/// field-merged across writers. A row therefore always describes a single
-/// position; a field a writer can't supply (a Kobo bookmark with no
-/// derivable CFI, a web CFI with no span) stays NULL as "not known", and
-/// the Kobo sync-out derives the missing half on demand rather than
-/// trusting a stale leftover.
-///
-/// One backstop narrows that rule: an **audio** write naming no
-/// `book_file_id`, landing on a row that names one, for a book carrying
-/// more than one audio file, is rejected wholly — same shape as a
-/// stale-clock rejection, the re-read returns the surviving row. Audio
-/// seconds are only meaningful within one file, so accepting such a write
-/// would turn the row into "N seconds within an unknown file" and destroy a
-/// position another client (e.g. iOS) recorded correctly (#1888). On a
-/// single-file book a fileless write still applies — the server resolves
-/// the same default file the manifest serves, so nothing is ambiguous.
-pub async fn upsert_progress(
-    pool: &SqlitePool,
-    user_id: i64,
-    update: &ProgressUpdate,
-) -> Result<ProgressRecord, ProgressError> {
-    // BEGIN IMMEDIATE avoids a stale-snapshot 517 on concurrent writes (#1862).
-    let mut tx = pool.begin_with("BEGIN IMMEDIATE").await?;
-    let record = upsert_progress_tx(&mut tx, user_id, update).await?;
-    tx.commit().await?;
-    Ok(record)
-}
-
-/// Transaction-scoped [`upsert_progress`], for callers that batch it with
-/// other writes (e.g. the Kobo `put_state` handler) inside one shared
-/// `Transaction` so a mid-batch failure rolls back every entry, not just the
-/// one that failed. The caller is responsible for committing or rolling back.
-pub async fn upsert_progress_tx(
-    tx: &mut Transaction<'_, Sqlite>,
-    user_id: i64,
-    update: &ProgressUpdate,
-) -> Result<ProgressRecord, ProgressError> {
-    let book_uuid = resolve_canonical_book_uuid_exec(&mut **tx, &update.book_uuid)
-        .await?
-        .ok_or(ProgressError::BookNotFound)?;
-    let fmt = format_str(update.format);
-
-    // Snapshot the row this write would land on, purely for the two WARNs
-    // this function logs (#1861) — neither reads back into the upsert itself.
-    let existing = existing_progress_snapshot(tx, user_id, &book_uuid, fmt).await?;
-    warn_if_rejected_by_timestamp_guard(&book_uuid, update.client_updated_at, existing.0);
-
-    // `strftime('%s','now')` returns TEXT; SQLite's default storage-class
-    // sort order ranks every INTEGER below every TEXT value regardless of
-    // magnitude, so `MIN(<int param>, <text now>)` would always pick the
-    // (unclamped) integer side without an explicit CAST here.
-    sqlx::query(
-        "INSERT INTO reading_progress
-            (user_id, book_uuid, format, epub_cfi, audio_position_seconds,
-             progress_percent, kobo_location, book_file_id,
-             updated_at, client_updated_at)
-         VALUES (?, ?, ?, ?, ?, ?, ?, ?, strftime('%s','now'),
-             MIN(
-                 COALESCE(?, CAST(strftime('%s','now') AS INTEGER)),
-                 CAST(strftime('%s','now') AS INTEGER)
-             ))
-         ON CONFLICT(user_id, book_uuid, format) DO UPDATE SET
-             epub_cfi = excluded.epub_cfi,
-             audio_position_seconds = excluded.audio_position_seconds,
-             progress_percent = excluded.progress_percent,
-             kobo_location = excluded.kobo_location,
-             book_file_id = excluded.book_file_id,
-             updated_at = strftime('%s','now'),
-             client_updated_at = excluded.client_updated_at
-         WHERE excluded.client_updated_at >=
-             COALESCE(reading_progress.client_updated_at, reading_progress.updated_at)
-           AND NOT (
-             excluded.format = 'audio'
-             AND excluded.book_file_id IS NULL
-             AND reading_progress.book_file_id IS NOT NULL
-             AND (SELECT COUNT(*) FROM book_files bf
-                    JOIN books b ON b.id = bf.book_id
-                   WHERE b.uuid = excluded.book_uuid
-                     AND bf.format IN ('M4B', 'M4A', 'MP3')) > 1
-           )",
-    )
-    .bind(user_id)
-    .bind(&book_uuid)
-    .bind(fmt)
-    // Blank-to-NULL at the bind, not just at `validate` — the row CHECK
-    // treats any non-NULL as a real position, so a whitespace CFI reaching
-    // an internal caller that skipped validation would store as an anchor.
-    .bind(update.epub_cfi.as_deref().filter(|s| !s.trim().is_empty()))
-    .bind(update.audio_position_seconds)
-    .bind(update.progress_percent)
-    .bind(
-        update
-            .kobo_location
-            .as_deref()
-            .filter(|s| !s.trim().is_empty()),
-    )
-    .bind(update.book_file_id)
-    .bind(update.client_updated_at)
-    .execute(&mut **tx)
-    .await?;
-
-    let row = sqlx::query(
-        "SELECT epub_cfi, audio_position_seconds, progress_percent, kobo_location,
-                book_file_id, updated_at,
-                COALESCE(client_updated_at, updated_at) AS client_updated_at
-         FROM reading_progress
-         WHERE user_id = ? AND book_uuid = ? AND format = ?",
-    )
-    .bind(user_id)
-    .bind(&book_uuid)
-    .bind(fmt)
-    .fetch_one(&mut **tx)
-    .await?;
-    let record = ProgressRecord {
-        book_uuid,
-        format: update.format,
-        epub_cfi: row.try_get::<Option<String>, _>("epub_cfi")?,
-        audio_position_seconds: row.try_get::<Option<f64>, _>("audio_position_seconds")?,
-        progress_percent: row.try_get::<Option<i64>, _>("progress_percent")?,
-        kobo_location: row.try_get::<Option<String>, _>("kobo_location")?,
-        book_file_id: row.try_get::<Option<i64>, _>("book_file_id")?,
-        updated_at: row.try_get::<i64, _>("updated_at")?,
-        client_updated_at: row.try_get::<i64, _>("client_updated_at")?,
-    };
-
-    // If this write was rejected (by either guard above), the row is
-    // unchanged, so `new` equals the pre-write snapshot and this can't
-    // fire — only a genuinely *accepted* jump ever crosses the threshold.
-    if update.format == ProgressFormat::Audio {
-        warn_if_audio_jumped_backward(&record, existing.1);
-    }
-
-    Ok(record)
-}
-
-/// The `(client_updated_at, audio_position_seconds)` a `(user, book, format)`
-/// row held immediately before an upsert — `None`/`None` when no row exists
-/// yet. Read-only; used solely to detect the two WARN conditions below.
-async fn existing_progress_snapshot(
-    tx: &mut Transaction<'_, Sqlite>,
-    user_id: i64,
-    book_uuid: &str,
-    fmt: &str,
-) -> Result<(Option<i64>, Option<f64>), ProgressError> {
-    let Some(row) = sqlx::query(
-        "SELECT COALESCE(client_updated_at, updated_at) AS client_updated_at,
-                audio_position_seconds
-         FROM reading_progress
-         WHERE user_id = ? AND book_uuid = ? AND format = ?",
-    )
-    .bind(user_id)
-    .bind(book_uuid)
-    .bind(fmt)
-    .fetch_optional(&mut **tx)
-    .await?
-    else {
-        return Ok((None, None));
-    };
-    Ok((
-        Some(row.try_get::<i64, _>("client_updated_at")?),
-        row.try_get::<Option<f64>, _>("audio_position_seconds")?,
-    ))
-}
-
-/// AC1 (#1861): a write is about to lose to the timestamp guard when the
-/// offered stamp — clamped forward to server-now, same as the SQL — is
-/// older than the stamp already on the row. Predicted here in Rust rather
-/// than read back from the upsert's own `WHERE`, purely for logging; a
-/// prediction that ever disagreed with the SQL would still change nothing,
-/// since the guard itself is unmodified.
-fn warn_if_rejected_by_timestamp_guard(
-    book_uuid: &str,
-    client_updated_at: Option<i64>,
-    stored_client_updated_at: Option<i64>,
-) {
-    let Some(stored_ts) = stored_client_updated_at else {
-        return;
-    };
-    let now = crate::auth::now_unix();
-    let offered = client_updated_at.unwrap_or(now).min(now);
-    if offered < stored_ts {
-        tracing::warn!(
-            book_uuid,
-            stored_client_updated_at = stored_ts,
-            offered_client_updated_at = offered,
-            "progress write rejected by timestamp guard"
-        );
-    }
-}
-
-/// AC2 (#1861): an *accepted* audio write whose new position lands more
-/// than [`AUDIO_BACKWARD_JUMP_THRESHOLD_SECONDS`] behind the old one — a
-/// rejected write leaves `record.audio_position_seconds` equal to the old
-/// value, so this can only fire for a write that actually landed.
-fn warn_if_audio_jumped_backward(
-    record: &ProgressRecord,
-    stored_audio_position_seconds: Option<f64>,
-) {
-    let (Some(old), Some(new)) = (stored_audio_position_seconds, record.audio_position_seconds)
-    else {
-        return;
-    };
-    if old - new > AUDIO_BACKWARD_JUMP_THRESHOLD_SECONDS {
-        tracing::warn!(
-            book_uuid = %record.book_uuid,
-            old_position_seconds = old,
-            new_position_seconds = new,
-            client_updated_at = record.client_updated_at,
-            "accepted audio write moved position backward past threshold"
-        );
-    }
-}
-
-/// Attach a server-derived KoboSpan location (and percent, when the row
-/// has none) to an existing epub row WITHOUT advancing any freshness
-/// clock — the derived half describes the same position the row already
-/// holds, so bumping `updated_at`/`client_updated_at` would re-fire the
-/// Kobo sync delta forever. Optimistic: no-ops unless the row still has
-/// no location and its event time is unchanged since the caller read it.
-/// Returns whether a row was updated.
-pub async fn attach_derived_kobo_location(
-    pool: &SqlitePool,
-    user_id: i64,
-    book_uuid: &str,
-    location_json: &str,
-    percent: Option<i64>,
-    expected_client_updated_at: i64,
-) -> Result<bool, ProgressError> {
-    let book_uuid = resolve_canonical_book_uuid(pool, book_uuid)
-        .await?
-        .ok_or(ProgressError::BookNotFound)?;
-    let result = sqlx::query(
-        "UPDATE reading_progress
-            SET kobo_location = ?,
-                progress_percent = COALESCE(progress_percent, ?)
-          WHERE user_id = ? AND book_uuid = ? AND format = 'epub'
-            AND kobo_location IS NULL
-            AND COALESCE(client_updated_at, updated_at) = ?",
-    )
-    .bind(location_json)
-    .bind(percent)
-    .bind(user_id)
-    .bind(&book_uuid)
-    .bind(expected_client_updated_at)
-    .execute(pool)
-    .await?;
-    Ok(result.rows_affected() > 0)
-}
-
-/// The device's own `Statistics` counters, mirrored so sync-out can hand
-/// them back. Never derived, never aggregated — see
-/// [`set_kobo_statistics_tx`].
-#[derive(Debug, Clone, Default, PartialEq, Eq)]
-pub struct KoboStatistics {
-    pub spent_reading_minutes: Option<i64>,
-    pub remaining_time_minutes: Option<i64>,
-    /// The device's `Statistics.LastModified`, clamped forward to server-now.
-    /// `None` when the device sent counters but no parseable clock.
-    pub updated_at: Option<i64>,
-}
-
-impl KoboStatistics {
-    /// Whether both counters are absent — nothing worth storing or echoing.
-    pub fn is_empty(&self) -> bool {
-        self.spent_reading_minutes.is_none() && self.remaining_time_minutes.is_none()
-    }
-}
-
-/// Mirror a device's `Statistics` block onto its epub position row (#1653);
-/// returns whether a row was updated.
-///
-/// UPDATE, not upsert: statistics annotate a position, they don't create one.
-/// A stamped write wins over an older or unstamped stored block (an unstamped
-/// one can't be echoed anyway); an unstamped write only takes an empty slot.
-/// So a second Kobo can't overwrite newer totals and have them echoed back as
-/// truth. The stamp clamps forward to server-now like [`upsert_progress_tx`]'s.
-pub async fn set_kobo_statistics_tx(
-    tx: &mut Transaction<'_, Sqlite>,
-    user_id: i64,
-    book_uuid: &str,
-    stats: &KoboStatistics,
-) -> Result<bool, ProgressError> {
-    let book_uuid = resolve_canonical_book_uuid_exec(&mut **tx, book_uuid)
-        .await?
-        .ok_or(ProgressError::BookNotFound)?;
-    // A NULL stamp stays NULL rather than becoming now: an invented clock
-    // would win the device's arbitration and overwrite its own totals.
-    let result = sqlx::query(
-        "UPDATE reading_progress
-            SET kobo_spent_reading_minutes = ?,
-                kobo_remaining_time_minutes = ?,
-                kobo_statistics_updated_at = CASE
-                    WHEN ? IS NULL THEN NULL
-                    ELSE MIN(?, CAST(strftime('%s','now') AS INTEGER))
-                END
-          WHERE user_id = ? AND book_uuid = ? AND format = 'epub'
-            AND (
-                (kobo_spent_reading_minutes IS NULL
-                 AND kobo_remaining_time_minutes IS NULL
-                 AND kobo_statistics_updated_at IS NULL)
-             OR (? IS NOT NULL AND ? >= COALESCE(kobo_statistics_updated_at, 0))
-            )",
-    )
-    .bind(stats.spent_reading_minutes)
-    .bind(stats.remaining_time_minutes)
-    .bind(stats.updated_at)
-    .bind(stats.updated_at)
-    .bind(user_id)
-    .bind(&book_uuid)
-    .bind(stats.updated_at)
-    .bind(stats.updated_at)
-    .execute(&mut **tx)
-    .await?;
-    Ok(result.rows_affected() > 0)
-}
-
-/// Fetch the current position row for `(user, book_uuid, format)`. Returns
-/// `Ok(None)` for an unknown book uuid or for a book with no row for that
-/// format yet.
-pub async fn get_progress(
-    pool: &SqlitePool,
-    user_id: i64,
-    book_uuid: &str,
-    format: ProgressFormat,
-) -> Result<Option<ProgressRecord>, ProgressError> {
-    let Some(canonical) = resolve_canonical_book_uuid(pool, book_uuid).await? else {
-        return Ok(None);
-    };
-    let fmt = format_str(format);
-    let Some(row) = sqlx::query(
-        "SELECT format, epub_cfi, audio_position_seconds, progress_percent, kobo_location,
-                book_file_id, updated_at,
-                COALESCE(client_updated_at, updated_at) AS client_updated_at
-         FROM reading_progress
-         WHERE user_id = ? AND book_uuid = ? AND format = ?",
-    )
-    .bind(user_id)
-    .bind(&canonical)
-    .bind(fmt)
-    .fetch_optional(pool)
-    .await?
-    else {
-        return Ok(None);
-    };
-    Ok(Some(ProgressRecord {
-        book_uuid: canonical,
-        format: parse_format(row.try_get::<String, _>("format")?.as_str()),
-        epub_cfi: row.try_get::<Option<String>, _>("epub_cfi")?,
-        audio_position_seconds: row.try_get::<Option<f64>, _>("audio_position_seconds")?,
-        progress_percent: row.try_get::<Option<i64>, _>("progress_percent")?,
-        kobo_location: row.try_get::<Option<String>, _>("kobo_location")?,
-        book_file_id: row.try_get::<Option<i64>, _>("book_file_id")?,
-        updated_at: row.try_get::<i64, _>("updated_at")?,
-        client_updated_at: row.try_get::<i64, _>("client_updated_at")?,
-    }))
-}
-
-/// Upsert the playback rate for `(user, book)` and return the saved preference.
-pub async fn set_playback_rate(
-    pool: &SqlitePool,
-    user_id: i64,
-    book_uuid: &str,
-    update: &AudiobookPlaybackRateUpdate,
-) -> Result<AudiobookPlaybackRateRecord, ProgressError> {
-    let canonical = resolve_canonical_book_uuid(pool, book_uuid)
-        .await?
-        .ok_or(ProgressError::BookNotFound)?;
-    let row = sqlx::query(
-        "INSERT INTO audiobook_playback_preferences
-            (user_id, book_uuid, playback_rate, updated_at)
-         VALUES (?, ?, ?, strftime('%s', 'now'))
-         ON CONFLICT(user_id, book_uuid) DO UPDATE SET
-            playback_rate = excluded.playback_rate,
-            updated_at = strftime('%s', 'now')
-         RETURNING playback_rate, updated_at",
-    )
-    .bind(user_id)
-    .bind(&canonical)
-    .bind(update.playback_rate)
-    .fetch_one(pool)
-    .await?;
-    Ok(AudiobookPlaybackRateRecord {
-        book_uuid: canonical,
-        playback_rate: row.try_get("playback_rate")?,
-        updated_at: row.try_get("updated_at")?,
-    })
-}
-
-/// Fetch the user's playback-rate preference for a book, if one exists.
-pub async fn get_playback_rate(
-    pool: &SqlitePool,
-    user_id: i64,
-    book_uuid: &str,
-) -> Result<Option<AudiobookPlaybackRateRecord>, ProgressError> {
-    let canonical = resolve_canonical_book_uuid(pool, book_uuid)
-        .await?
-        .ok_or(ProgressError::BookNotFound)?;
-    let Some(row) = sqlx::query(
-        "SELECT playback_rate, updated_at
-         FROM audiobook_playback_preferences
-         WHERE user_id = ? AND book_uuid = ?",
-    )
-    .bind(user_id)
-    .bind(&canonical)
-    .fetch_optional(pool)
-    .await?
-    else {
-        return Ok(None);
-    };
-    Ok(Some(AudiobookPlaybackRateRecord {
-        book_uuid: canonical,
-        playback_rate: row.try_get("playback_rate")?,
-        updated_at: row.try_get("updated_at")?,
-    }))
-}
-
-/// The user's most recent progress rows across both formats, newest first
-/// by **client event time** — `COALESCE(client_updated_at, updated_at)`, so
-/// a queued offline replay landing late doesn't jump a week-old book to the
-/// top. Feeds the "pick up where you left off" surfaces via
-/// [`resume_points`].
-///
-/// Books the user has marked `unread` or `finished` are excluded: both are
-/// statements that they are not mid-book, and a completed book otherwise sat
-/// at the top of the rail until five newer positions pushed it off.
-///
-/// A **missing** `book_read_status` row keeps its position. 0046 treats
-/// absence as `unread` everywhere else, and this is the deliberate exception:
-/// absence here means "nothing has been said", not "not reading". Every
-/// status write is best-effort (`read_status_auto`, the players), so hiding on
-/// absence would drop a book off the rail whenever one of those requests was
-/// lost — and would hide every row written before that auto-transition
-/// existed.
-///
-/// The join is per-book while progress is per-`(book, format)`, so finishing
-/// a dual-format book clears both its reading and its listening card. Read
-/// status is a fact about the book, not about one of its files.
-pub async fn recent_progress(
-    pool: &SqlitePool,
-    user_id: i64,
-    limit: i64,
-) -> Result<Vec<ProgressRecord>, ProgressError> {
-    let rows = sqlx::query(
-        "SELECT rp.book_uuid, rp.format, rp.epub_cfi, rp.audio_position_seconds,
-                rp.progress_percent, rp.kobo_location, rp.book_file_id,
-                rp.updated_at,
-                COALESCE(rp.client_updated_at, rp.updated_at) AS client_updated_at
-         FROM reading_progress rp
-         LEFT JOIN book_read_status rs
-                ON rs.user_id = rp.user_id AND rs.book_uuid = rp.book_uuid
-         WHERE rp.user_id = ?
-           AND (rs.status IS NULL OR rs.status = 'reading')
-         ORDER BY COALESCE(rp.client_updated_at, rp.updated_at) DESC, rp.book_uuid
-         LIMIT ?",
-    )
-    .bind(user_id)
-    .bind(limit)
-    .fetch_all(pool)
-    .await?;
-    rows.into_iter()
-        .map(|row| {
-            Ok(ProgressRecord {
-                book_uuid: row.try_get::<String, _>("book_uuid")?,
-                format: parse_format(row.try_get::<String, _>("format")?.as_str()),
-                epub_cfi: row.try_get::<Option<String>, _>("epub_cfi")?,
-                audio_position_seconds: row.try_get::<Option<f64>, _>("audio_position_seconds")?,
-                progress_percent: row.try_get::<Option<i64>, _>("progress_percent")?,
-                kobo_location: row.try_get::<Option<String>, _>("kobo_location")?,
-                book_file_id: row.try_get::<Option<i64>, _>("book_file_id")?,
-                updated_at: row.try_get::<i64, _>("updated_at")?,
-                client_updated_at: row.try_get::<i64, _>("client_updated_at")?,
-            })
-        })
-        .collect()
-}
-
-/// [`recent_progress`] joined with book metadata and, for audio rows, the
-/// whole-book duration + chapter position. Rows whose book has since been
-/// deleted (progress soft-references `books.uuid` — no cascade) are skipped
-/// rather than erroring, so one ghosted book can't blank the landing card.
-pub async fn resume_points(
-    pool: &SqlitePool,
-    user_id: i64,
-    limit: i64,
-) -> Result<Vec<ResumePoint>, ProgressError> {
-    let records = recent_progress(pool, user_id, limit).await?;
-    let mut points = Vec::with_capacity(records.len());
-    for mut record in records {
-        let Some(book) = crate::get_book_by_uuid(pool, &record.book_uuid).await? else {
-            continue;
-        };
-        let audio = match record.format {
-            ProgressFormat::Audio => audio_totals(pool, &record.book_uuid, &record).await?,
-            ProgressFormat::Epub => None,
-        };
-        let (total_duration_seconds, chapter_number, chapter_count) = match audio {
-            Some(totals) => {
-                // Overwrite rather than trust the stored id: it may name a
-                // `book_files` row the reindex has since replaced, and the
-                // Continue CTA links straight at `?file_id=` — a dead id
-                // would open the player on a manifest that 404s.
-                record.book_file_id = Some(totals.book_file_id);
-                (
-                    Some(totals.total_duration_seconds),
-                    totals.chapter_number,
-                    totals.chapter_count,
-                )
-            }
-            None => {
-                // Audio row whose book has no audio file left (and every
-                // epub row): drop the stored id rather than hand a CTA an
-                // id that resolves to nothing.
-                record.book_file_id = None;
-                (None, None, None)
-            }
-        };
-        // Only rows that will render a listening card need the preference —
-        // it feeds the rate-adjusted "left" readout on the resume surfaces.
-        // Direct lookup, not `get_playback_rate`: `record.book_uuid` was
-        // canonicalized by `upsert_progress_tx` at write time, so re-resolving
-        // per row would double this endpoint's query count for a no-op. A row
-        // predating a later merge can miss the preference and fall back to 1x
-        // until its next write re-canonicalizes it — cosmetic, and accepted.
-        let playback_rate = match total_duration_seconds {
-            Some(_) => sqlx::query(
-                "SELECT playback_rate FROM audiobook_playback_preferences
-                 WHERE user_id = ? AND book_uuid = ?",
-            )
-            .bind(user_id)
-            .bind(&record.book_uuid)
-            .fetch_optional(pool)
-            .await?
-            .map(|row| row.try_get::<f64, _>("playback_rate"))
-            .transpose()?,
-            None => None,
-        };
-        points.push(ResumePoint {
-            record,
-            book,
-            total_duration_seconds,
-            chapter_number,
-            chapter_count,
-            playback_rate,
-        });
-    }
-    Ok(points)
-}
-
-/// Which audio file a resume point plays, plus the duration and chapter
-/// position measured against **that** file.
-struct AudioTotals {
-    book_file_id: i64,
-    total_duration_seconds: f64,
-    chapter_number: Option<i64>,
-    chapter_count: Option<i64>,
-}
-
-/// Resolve the audio file for a progress row and measure duration + chapter
-/// position against it. `None` when the book has no resolvable audio file
-/// (e.g. every file was removed after the position was saved).
-///
-/// The row's stored `book_file_id` picks the file for a book carrying more
-/// than one audiobook, so the resume card reads out the narration the user
-/// was actually in. It is a soft reference (rule 06) — a stale id, or one
-/// belonging to another book, falls back to the first audio file by ordinal,
-/// which is what the whole feature did before the id was recorded.
-async fn audio_totals(
-    pool: &SqlitePool,
-    uuid: &str,
-    record: &ProgressRecord,
-) -> Result<Option<AudioTotals>, ProgressError> {
-    let stored = match record.book_file_id {
-        Some(id) => hls::resolve_audiobook_file(pool, uuid, Some(id)).await?,
-        None => None,
-    };
-    let resolved = match stored {
-        Some(resolved) => resolved,
-        None => match hls::resolve_audiobook(pool, uuid).await? {
-            Some(resolved) => resolved,
-            None => return Ok(None),
-        },
-    };
-    let parts = hls::get_parts(pool, resolved.book_file_id).await?;
-    let total: f64 = parts.iter().map(|p| p.duration_seconds).sum();
-    let mut chapters = hls::get_chapters(pool, resolved.book_file_id).await?;
-    chapters.sort_by(|a, b| a.start_seconds.total_cmp(&b.start_seconds));
-    let position = record.audio_position_seconds.unwrap_or(0.0);
-    Ok(Some(AudioTotals {
-        book_file_id: resolved.book_file_id,
-        total_duration_seconds: total,
-        chapter_number: chapter_number_at(&chapters, position),
-        chapter_count: (!chapters.is_empty()).then_some(chapters.len() as i64),
-    }))
-}
-
-/// 1-based chapter number at `elapsed` seconds, mirroring the player's
-/// index-plus-one display (not the stored `file_chapters.ordinal`, which is
-/// container-supplied and not guaranteed dense).
-fn chapter_number_at(chapters: &[ChapterInfo], elapsed: f64) -> Option<i64> {
-    if chapters.is_empty() {
-        return None;
-    }
-    let idx = chapters
-        .partition_point(|c| c.start_seconds <= elapsed)
-        .saturating_sub(1);
-    Some(idx as i64 + 1)
-}
-
-/// Append one session row inside an existing transaction. Returns `Ok(true)`
-/// when a row was inserted and `Ok(false)` when the report was skipped
-/// because the `book_uuid` resolves to neither a `books` row nor a
-/// `merged_uuids` entry (best-effort telemetry — a session that outlived its
-/// book is not an integrity failure). A format-merged or auto-attached uuid
-/// resolves to the surviving book and is recorded.
-///
-/// The caller is responsible for committing or rolling back the transaction.
-/// Batch writers that already pre-resolved every uuid via
-/// [`crate::resolve_canonical_book_uuids_bulk_exec`] should skip this wrapper
-/// and call [`insert_session_tx`] directly to avoid the per-row SELECT.
-pub async fn record_session_tx(
-    tx: &mut Transaction<'_, Sqlite>,
-    user_id: i64,
-    report: &SessionReport,
-) -> Result<bool, ProgressError> {
-    // Resolve through the same merged-uuid-aware path as `upsert_progress`,
-    // so a uuid that was format-merged or auto-attached after the session
-    // started still records against the surviving book instead of being
-    // silently dropped. `Ok(false)` now means "unknown in neither `books`
-    // nor `merged_uuids`".
-    let Some(book_uuid) = resolve_canonical_book_uuid_exec(&mut **tx, &report.book_uuid).await?
-    else {
-        return Ok(false);
-    };
-    insert_session_tx(tx, user_id, report, &book_uuid).await?;
-    Ok(true)
-}
-
-/// Insert one session row into the correct per-format table using a
-/// **pre-resolved** canonical `books.uuid`. This is the INSERT-only half of
-/// [`record_session_tx`], exposed so batch writers (see `post_sessions`) can
-/// pre-resolve every uuid in the batch through
-/// [`crate::resolve_canonical_book_uuids_bulk_exec`] and then loop through
-/// pure inserts — collapsing an N-report batch's 2N queries into
-/// `chunks + N`. The caller is responsible for committing or rolling back
-/// the transaction; the caller also owns the "skip on unknown uuid" branch
-/// (an entry missing from the bulk map).
-pub async fn insert_session_tx(
-    tx: &mut Transaction<'_, Sqlite>,
-    user_id: i64,
-    report: &SessionReport,
-    canonical_uuid: &str,
-) -> Result<(), ProgressError> {
-    // OR IGNORE against the partial-unique `(user_id, client_id)` index from
-    // migration 0052: a report the client replayed because it never saw the
-    // reply collapses onto the row it already wrote instead of doubling the
-    // reading time it represents. Reports without a client id are
-    // unconstrained and insert as before.
-    match report.format {
-        ProgressFormat::Epub => {
-            sqlx::query(
-                "INSERT OR IGNORE INTO reading_sessions
-                    (user_id, book_uuid, started_at, ended_at, seconds_read, device_id, client_id)
-                 VALUES (?, ?, ?, ?, ?, ?, ?)",
-            )
-            .bind(user_id)
-            .bind(canonical_uuid)
-            .bind(report.started_at)
-            .bind(report.ended_at)
-            .bind(report.progress_units)
-            .bind(report.device_id)
-            .bind(report.client_id.as_deref())
-            .execute(&mut **tx)
-            .await?;
-        }
-        ProgressFormat::Audio => {
-            sqlx::query(
-                "INSERT OR IGNORE INTO listening_sessions
-                    (user_id, book_uuid, started_at, ended_at, seconds_listened, device_id, client_id)
-                 VALUES (?, ?, ?, ?, ?, ?, ?)",
-            )
-            .bind(user_id)
-            .bind(canonical_uuid)
-            .bind(report.started_at)
-            .bind(report.ended_at)
-            .bind(report.progress_units)
-            .bind(report.device_id)
-            .bind(report.client_id.as_deref())
-            .execute(&mut **tx)
-            .await?;
-        }
-    }
-    Ok(())
-}
-
-/// Append one session row to the per-format table. Returns `Ok(true)` when
-/// a row was inserted and `Ok(false)` when the report was skipped because
-/// the `book_uuid` is unknown. The handler surfaces the inserted count to
-/// the client so it can tell which queued reports actually persisted.
-///
-/// For batch inserts, prefer the pattern in `post_sessions`: pre-resolve
-/// every uuid via [`crate::resolve_canonical_book_uuids_bulk_exec`], then
-/// loop [`insert_session_tx`] inside a caller-managed transaction so the
-/// entire batch rolls back atomically on error and no per-row SELECT fires.
-pub async fn record_session(
-    pool: &SqlitePool,
-    user_id: i64,
-    report: &SessionReport,
-) -> Result<bool, ProgressError> {
-    // Same BEGIN IMMEDIATE reasoning as `upsert_progress` above (#1862).
-    let mut tx = pool.begin_with("BEGIN IMMEDIATE").await?;
-    let result = record_session_tx(&mut tx, user_id, report).await?;
-    tx.commit().await?;
-    Ok(result)
-}
-
-#[cfg(test)]
-mod tests;
